Implement reusable ProjectCard component for consistency
- Create shared ProjectCard component to ensure identical project cards across pages
- Update FeaturedProjects to use ProjectCard component
- Update Projects page to use ProjectCard component
- Remove unused Link import from Projects page
- Fix TypeScript skill category type to include 'languages' and 'tools'
