+++ app/projects/page.tsx
@@ -1,10 +1,9 @@
 "use client"
 
 import { useState } from "react"
-import Link from "next/link"
-import { ArrowUpRight, Github, BookOpen } from "lucide-react"
 import { Project } from "@/types"
 import { cn } from "@/lib/utils"
+import { ProjectCard } from "@/components/ProjectCard"
 
 const allProjects: Project[] = [
   {
@@ -209,222 +208,7 @@ export default function ProjectsPage() {
 
         <div className="mx-auto mt-8 sm:mt-12 lg:mt-16 grid max-w-2xl grid-cols-1 gap-4 sm:gap-6 lg:gap-8 lg:max-w-none lg:grid-cols-2 xl:grid-cols-3">
           {filteredProjects.map((project) => (
-            <article
-              key={project.id}
-              className="group relative flex flex-col overflow-hidden rounded-lg border bg-card shadow-sm transition-all hover:shadow-lg"
-            >
-              <div className="aspect-[4/3] sm:aspect-[16/9] w-full overflow-hidden bg-gradient-to-br from-primary/10 via-accent/5 to-muted relative">
-                <div className="absolute inset-0 bg-gradient-to-t from-black/20 to-transparent" />
-                
-                {/* Animated project visualizations matching FeaturedProjects */}
-                <div className="absolute inset-0 opacity-50">
-                  {project.id === "1" && (
-                    <div className="w-full h-full bg-gradient-to-br from-blue-600/60 via-cyan-500/50 to-teal-400/40 flex items-center justify-center overflow-hidden">
-                      <div className="grid grid-cols-12 gap-1 opacity-80 rotate-3 animate-pulse">
-                        {[...Array(60)].map((_, i) => (
-                          <div key={i} className={`w-1.5 bg-gradient-to-t from-blue-500 to-cyan-400 rounded-sm animate-bounce`} 
-                               style={{
-                                 height: `${Math.floor(Math.random() * 50) + 15}px`,
-                                 animationDelay: `${(i * 0.1) % 3}s`,
-                                 animationDuration: `${1 + (i % 3) * 0.5}s`
-                               }} />
-                        ))}
-                      </div>
-                      <div className="absolute inset-0">
-                        {[...Array(5)].map((_, i) => (
-                          <div key={i} className="absolute h-0.5 w-full bg-gradient-to-r from-transparent via-cyan-400/60 to-transparent"
-                               style={{
-                                 top: `${20 + i * 15}%`,
-                                 animation: `data-stream ${2 + i * 0.5}s linear infinite`,
-                                 animationDelay: `${i * 0.4}s`
-                               }} />
-                        ))}
-                      </div>
-                    </div>
-                  )}
-                  {project.id === "2" && (
-                    <div className="w-full h-full bg-gradient-to-br from-purple-600/60 via-pink-500/50 to-violet-400/40 flex items-center justify-center">
-                      <div className="relative">
-                        <div className="w-32 h-32 border-4 border-purple-300 rounded-2xl opacity-70 rotate-12 animate-pulse" />
-                        <div className="absolute top-4 left-4 w-8 h-8 bg-gradient-to-r from-purple-400 to-pink-400 rounded-lg opacity-90 animate-bounce" />
-                        <div className="absolute bottom-4 right-4 w-6 h-6 bg-gradient-to-r from-pink-400 to-violet-400 rounded-full opacity-90 animate-bounce" style={{animationDelay: '0.5s'}} />
-                        <div className="absolute top-8 right-8 w-4 h-4 bg-gradient-to-r from-violet-400 to-purple-400 rounded opacity-80 animate-ping" />
-                      </div>
-                    </div>
-                  )}
-                  {(project.id === "3" || project.id === "6" || project.category === "automation") && (
-                    <div className="w-full h-full bg-gradient-to-br from-green-600/60 via-emerald-500/50 to-teal-400/40 flex items-center justify-center overflow-hidden">
-                      <div className="grid grid-cols-8 gap-3 opacity-70 -rotate-6">
-                        {[...Array(32)].map((_, i) => (
-                          <div key={i} className="w-4 h-4 bg-gradient-to-br from-green-400 to-emerald-400 rounded-full animate-bounce" 
-                               style={{
-                                 animationDelay: `${i * 0.1}s`,
-                                 animationDuration: `${1.5 + (i % 4) * 0.3}s`
-                               }} />
-                        ))}
-                      </div>
-                      <div className="absolute inset-0">
-                        {[...Array(6)].map((_, i) => (
-                          <div key={i} className="absolute w-0.5 h-full bg-gradient-to-b from-transparent via-emerald-400/50 to-transparent"
-                               style={{
-                                 left: `${15 + i * 12}%`,
-                                 animation: `matrix-rain ${3 + i * 0.5}s linear infinite`,
-                                 animationDelay: `${i * 0.3}s`
-                               }} />
-                        ))}
-                      </div>
-                    </div>
-                  )}
-                  {project.category === "blockchain" && (
-                    <div className="w-full h-full bg-gradient-to-br from-indigo-600/60 via-purple-500/50 to-blue-400/40 flex items-center justify-center overflow-hidden">
-                      {/* Blockchain network visualization */}
-                      <div className="relative w-full h-full">
-                        {/* Network nodes */}
-                        {[...Array(12)].map((_, i) => (
-                          <div key={i} className="absolute w-3 h-3 bg-gradient-to-r from-indigo-400 to-purple-400 rounded-full animate-pulse"
-                               style={{
-                                 top: `${Math.random() * 80 + 10}%`,
-                                 left: `${Math.random() * 80 + 10}%`,
-                                 animationDelay: `${i * 0.2}s`,
-                                 animationDuration: `${2 + Math.random()}s`
-                               }} />
-                        ))}
-                        {/* Connecting lines */}
-                        {[...Array(8)].map((_, i) => (
-                          <div key={i} className="absolute h-0.5 bg-gradient-to-r from-transparent via-purple-400/60 to-transparent animate-pulse"
-                               style={{
-                                 top: `${20 + i * 8}%`,
-                                 left: `${10 + (i % 3) * 25}%`,
-                                 width: `${30 + Math.random() * 20}%`,
-                                 transform: `rotate(${Math.random() * 60 - 30}deg)`,
-                                 animationDelay: `${i * 0.3}s`
-                               }} />
-                        ))}
-                        {/* Central blockchain icon */}
-                        <div className="absolute top-1/2 left-1/2 transform -translate-x-1/2 -translate-y-1/2">
-                          <div className="w-8 h-8 border-2 border-indigo-300 rounded opacity-80 animate-bounce">
-                            <div className="w-full h-full bg-gradient-to-br from-indigo-400 to-purple-400 rounded animate-pulse" />
-                          </div>
-                        </div>
-                      </div>
-                    </div>
-                  )}
-                  {project.category === "ai" && project.id !== "2" && (
-                    <div className="w-full h-full bg-gradient-to-br from-violet-600/60 via-fuchsia-500/50 to-purple-400/40 flex items-center justify-center overflow-hidden">
-                      {/* AI neural network pattern */}
-                      <div className="relative w-full h-full">
-                        {/* Neural connections */}
-                        <div className="absolute inset-0 opacity-60">
-                          {[...Array(20)].map((_, i) => (
-                            <div key={i} className="absolute w-1 h-1 bg-violet-400 rounded-full animate-ping"
-                                 style={{
-                                   top: `${Math.random() * 100}%`,
-                                   left: `${Math.random() * 100}%`,
-                                   animationDelay: `${i * 0.1}s`,
-                                   animationDuration: `${1 + Math.random()}s`
-                                 }} />
-                          ))}
-                        </div>
-                        {/* Brain-like structure */}
-                        <div className="absolute top-1/2 left-1/2 transform -translate-x-1/2 -translate-y-1/2">
-                          <div className="relative">
-                            <div className="w-16 h-12 border-2 border-violet-300 rounded-full opacity-70 animate-pulse" />
-                            <div className="absolute top-1 left-2 w-3 h-3 bg-violet-400 rounded-full animate-bounce" />
-                            <div className="absolute top-2 right-3 w-2 h-2 bg-fuchsia-400 rounded-full animate-bounce" style={{animationDelay: '0.3s'}} />
-                            <div className="absolute bottom-1 left-4 w-2.5 h-2.5 bg-purple-400 rounded-full animate-bounce" style={{animationDelay: '0.6s'}} />
-                          </div>
-                        </div>
-                        {/* Flowing data streams */}
-                        {[...Array(4)].map((_, i) => (
-                          <div key={i} className="absolute w-full h-0.5 bg-gradient-to-r from-transparent via-violet-400/50 to-transparent"
-                               style={{
-                                 top: `${25 + i * 15}%`,
-                                 animation: `data-stream ${2.5 + i * 0.3}s ease-in-out infinite`,
-                                 animationDelay: `${i * 0.4}s`
-                               }} />
-                        ))}
-                      </div>
-                    </div>
-                  )}
-                </div>
-                
-                {/* Category badge moved to corner */}
-                <div className="absolute top-4 right-4 z-10">
-                  <div className="text-white/90 text-xs font-medium px-2.5 py-1 rounded-full bg-black/40 backdrop-blur-sm border border-white/20">
-                    {project.category === 'enterprise' ? 'Enterprise' : 
-                     project.category === 'ai' ? 'AI-Powered' : 
-                     project.category === 'automation' ? 'Automation' :
-                     project.category === 'blockchain' ? 'Blockchain' : 'Project'}
-                  </div>
-                </div>
-              </div>
-              <div className="flex flex-1 flex-col p-4 sm:p-5 lg:p-6">
-                <div className="flex-1">
-                  <div className="flex items-center justify-between">
-                    <span className="text-xs text-muted-foreground">
-                      {project.date}
-                    </span>
-                    {project.featured && (
-                      <span className="inline-flex items-center rounded-full bg-primary/10 px-2 py-1 text-xs font-medium text-primary">
-                        Featured
-                      </span>
-                    )}
-                  </div>
-                  <h3 className="mt-3 text-lg sm:text-xl font-semibold">
-                    {project.title}
-                  </h3>
-                  <p className="mt-3 text-sm sm:text-base text-muted-foreground line-clamp-2">{project.description}</p>
-                  <div className="mt-4 flex flex-wrap gap-2">
-                    {project.technologies.slice(0, 4).map((tech) => (
-                      <span
-                        key={tech}
-                        className="inline-flex items-center rounded-md bg-secondary px-2.5 py-0.5 text-xs font-medium text-secondary-foreground"
-                      >
-                        {tech}
-                      </span>
-                    ))}
-                    {project.technologies.length > 4 && (
-                      <span className="inline-flex items-center text-xs text-muted-foreground">
-                        +{project.technologies.length - 4} more
-                      </span>
-                    )}
-                  </div>
-                </div>
-                <div className="mt-4 sm:mt-6 flex items-center gap-3 sm:gap-4">
-                  {project.blogArticle && (
-                    <Link
-                      href={project.blogArticle}
-                      className="inline-flex items-center text-sm font-medium text-primary hover:text-primary/80 transition-colors min-h-[44px] py-2"
-                    >
-                      <BookOpen className="mr-1.5 h-4 w-4" />
-                      Read More
-                    </Link>
-                  )}
-                  {project.githubUrl && (
-                    <a
-                      href={project.githubUrl}
-                      target="_blank"
-                      rel="noopener noreferrer"
-                      className="inline-flex items-center text-sm font-medium text-muted-foreground hover:text-foreground transition-colors min-h-[44px] py-2"
-                    >
-                      <Github className="mr-1.5 h-4 w-4" />
-                      Code
-                    </a>
-                  )}
-                  {project.liveUrl && (
-                    <a
-                      href={project.liveUrl}
-                      target="_blank"
-                      rel="noopener noreferrer"
-                      className="inline-flex items-center text-sm font-medium text-primary hover:text-primary/80 transition-colors min-h-[44px] py-2"
-                    >
-                      Live Demo
-                      <ArrowUpRight className="ml-1 h-3.5 w-3.5" />
-                    </a>
-                  )}
-                </div>
-              </div>
-            </article>
+            <ProjectCard key={project.id} project={project} />
           ))}
         </div>
